amazing rebind fix
this fix, indeed, is a work of art, let me explain...The problem was displayed as this : the basic attack and the charged attack are binded to the same button, so they will be counted as duplicate bindings for the rebinding script. CRAP! HOW CAN I FIX THIS?? Then, i got an idea, a genious idea, an idea that not a lot of of people could have had, and i figured out that there is a list of interactions stored inside the binding that represents if the binding needs to be hold, just tapped, multi-tapped etc... so if i had a check on the duplicate binding to check not only the binding but also the interactions of the binding, i would get the perfect check.

This is my story,
Olivier Maurin